sex idk

+++ resources/shaders/shader.vert
@@ -2,7 +2,7 @@
 #extension GL_ARB_separate_shader_objects : enable
 
 layout(location = 0) in vec2 inPosition;
-layout(location = 1) in vec4 inColour;
+layout(location = 1) in uvec4 inColour;
 
 layout(location = 0) out vec4 fragColour;
